Refactor to collect register related js files
- register 페이지와 관련된 여러 js 파일을 하나의 js 파일에 모아서 호출하도록 구조 수정
